+++ r-network.tf
@@ -4,6 +4,7 @@ resource "azurerm_public_ip" "public_ip" {
   resource_group_name = var.resource_group_name
   allocation_method   = "Static"
   domain_name_label   = coalesce(var.custom_dns_label, local.vm_name)
+  sku                 = var.public_ip_sku
 
   tags = merge(local.default_tags, var.extra_tags)
 }